chore: Move cloudstream example to own package
