+++ springwolf-examples/springwolf-cloud-stream-example/src/main/java/io/github/stavshamir/springwolf/example/cloudstream/payload/ConsumerPayload.java
@@ -1,4 +1,4 @@
-package io.github.stavshamir.springwolf.example.payload;
+package io.github.stavshamir.springwolf.example.cloudstream.payload;
 
 public class ConsumerPayload {
