+++ springwolf-examples/springwolf-cloud-stream-example/src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.cloud.stream.bindings.consumerMethod-in-0.destination=consumer-method-inp
 spring.application.name=Springwolf example project - Cloud Stream
 
 springwolf.enabled=true
-springwolf.docket.base-package=io.github.stavshamir.springwolf.example
+springwolf.docket.base-package=io.github.stavshamir.springwolf.example.cloudstream
 springwolf.docket.info.title=${spring.application.name}
 springwolf.docket.info.version=1.0.0
 springwolf.docket.info.description=Springwolf example project to demonstrate springwolfs abilities
